some STG word size fixes

+++ runtime/stg.h
@@ -4,6 +4,7 @@
 #include <stddef.h>
 #include <assert.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 //------ stack and heap objects
 
@@ -50,7 +51,7 @@ typedef FnPtr (*CmmFnPtr)();
 typedef struct {
   ArgType argType;        // superfluous, for sanity checking
   union {
-    int i;
+    int64_t i;
     double d;
     float f;
     bool b;
@@ -155,15 +156,15 @@ extern void showStgVal(PtrOrLiteral);
 extern void showIT(InfoTab *);
 extern int getObjSize(Obj *);
 
-#define hibits (~0 << sizeof(int)/2 * 8)
-#define lobits (~0 & ~hibits)
+#define hibits (~0L << sizeof(uintptr_t)/2 * 8)
+#define lobits (~0L & ~hibits)
 
-#define PNPACK(pargc,nargc) ((pargc) | ((nargc) << (sizeof(int)/2 * 8)))
+#define PNPACK(pargc,nargc) ((pargc) | (((uintptr_t)(nargc)) << (sizeof(uintptr_t)/2 * 8)))
 
 #define PNUNPACK(n,pargc,nargc) \
   do { \
     pargc = (n) & lobits;				\
-    nargc = ((unsigned int) (n)) >> (sizeof(int)/2 * 8);	\
+    nargc = ((uintptr_t) (n)) >> (sizeof(uintptr_t)/2 * 8);	\
   } while (0)
 
 // allocate Obj on heap, returning pointer to new Obj
